Update/improve chat security

+++ chat/src/chat/message/message.service.ts
@@ -5,9 +5,12 @@ import { UpdateMessageDto } from './dto/update-message.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
-import { Message } from './message.entity';
+import { HttpException, HttpStatus } from '@nestjs/common';
+import { Logger } from '@nestjs/common';
 
+import { Message } from './message.entity';
 import { Channel } from 'src/chat/channel/channel.entity';
+import { User } from 'src/user/user.entity';
 
 @Injectable()
 export class MessageService {
@@ -16,21 +19,37 @@ export class MessageService {
     private readonly messagesRepository: Repository<Message>,
     @InjectRepository(Channel)
     private readonly channelsRepository: Repository<Channel>,
+	@InjectRepository(User)
+    private readonly usersRepository: Repository<User>,
   ) {}
 
-
-  // @Inject(ChannelService)
-  // private readonly channelService: ChannelService;
+  private logger: Logger = new Logger('MessageService');
 
   async create(data: CreateMessageDto): Promise<Message> 
   {
     const message = new Message();
-    message.text = data.text;
-    message.sender = data.sender;
+	if (data.text && data.text.length > 0)	
+		message.text = data.text;
+	else
+		throw new HttpException("message text is empty", HttpStatus.FAILED_DEPENDENCY);
 
-    const channel = await this.channelsRepository.findOneBy({id: data.channelId}) ;
-    message.channel = channel;
+	if (data.private)
+		message.private = data.private;
 
+	if (data.senderId)
+	{
+		const sender = await this.usersRepository.findOneBy({id: data.senderId});
+		if (!sender)
+			throw new HttpException("this user sender doesn't exist", HttpStatus.FAILED_DEPENDENCY);
+		message.sender = sender;
+	}
+	if (data.channelId)
+	{
+		const channel = await this.channelsRepository.findOneBy({id: data.channelId}) ;
+		if (!channel)
+			throw new HttpException("this channel doesn't exist", HttpStatus.FAILED_DEPENDENCY);
+		message.channel = channel;
+	}
     return this.messagesRepository.save(message);
   }
 
@@ -39,8 +58,11 @@ export class MessageService {
     return this.messagesRepository.find();
   }
 
-  findOne(id: number)
+  async findOne(id: number)
   {
+	const message = await this.messagesRepository.findOneBy({ id: id })
+	if (!message)
+		throw new HttpException('Message not found', HttpStatus.NOT_FOUND);
     return this.messagesRepository.findOneBy({ id: id });
   }
 
@@ -51,6 +73,13 @@ export class MessageService {
 
   async remove(id: number) 
   {
-    return this.messagesRepository.delete(id);
+	const message = await this.messagesRepository.findOneBy({ id: id })
+	if (!message)
+		throw new HttpException('Message not found', HttpStatus.NOT_FOUND);
+	else
+	{
+		this.messagesRepository.delete(id);
+		throw new HttpException(`message #${id} was deleted successfully`, HttpStatus.OK);
+	}
   }
 }
